+++ app/Http/Controllers/Backend/UserRoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\Services\Interfaces\UserRoleServiceInterface as UserRoleService;
 use App\Repositories\Interfaces\UserRoleRepositoryInterface as UserRoleRepository;
+use App\Repositories\Interfaces\PermissionRepositoryInterface as PermissionRepository;
 
 use App\Http\Requests\UpdateUserRequest;
 use App\Http\Requests\UserRoleRequest;
@@ -15,16 +16,20 @@ class UserRoleController extends Controller{
 
     protected $userRoleService;
     protected $userRoleRepository;
+    protected $permissionRepository;
 
     public function __construct(
         UserRoleService $userRoleService,
         UserRoleRepository $userRoleRepository,
+        PermissionRepository $permissionRepository,
     ) {
         $this->userRoleService = $userRoleService;
         $this->userRoleRepository = $userRoleRepository;
+        $this->permissionRepository = $permissionRepository;
     }
 
     public function index(Request $request){
+        $this->authorize('modules', 'user.role.index');
 
         $perPage = $request->integer('perpage', 10);
         $userRoles = $this->userRoleService->paginate($request, $perPage);
@@ -40,6 +45,8 @@ class UserRoleController extends Controller{
 
 
     public function create(){
+        $this->authorize('modules', 'user.role.create');
+
         $config['method'] = 'create';
         $config['seo'] = config('apps.userRole.create');
 
@@ -60,6 +67,8 @@ class UserRoleController extends Controller{
 
 
     public function edit($id){
+        $this->authorize('modules', 'user.role.edit');
+
         $userRoles = $this->userRoleRepository->findById($id);
 
         $config['method'] = 'edit';
@@ -72,6 +81,7 @@ class UserRoleController extends Controller{
         ));
     }
 
+
     public function update($id, UserRoleRequest $request){
         if ($this->userRoleService->update($id, $request)) {
             return redirect()->route('user.role.index')->with('success', 'Cập nhập nhóm thành viên thành công !');
@@ -79,11 +89,38 @@ class UserRoleController extends Controller{
         return redirect()->route('user.role.index')->with('error', 'Cập nhập nhóm thành viên thất bại !');
     }
 
+
     public function destroy($id){
+        $this->authorize('modules', 'user.role.destroy');
         if ($this->userRoleService->destroy($id)) {
             return redirect()->route('user.role.index')->with('success', 'Xóa nhóm thành viên thành công !');
         }
         return redirect()->route('user.role.index')->with('error', 'Xóa nhóm thành viên thất bại !');
+    }
+
+
+    public function permission(){
+        // $this->authorize('modules', 'user.role.permission');
+        $template = 'backend.user.role_user.permission';
+        $config['seo'] = config('apps.permission.index');
+        $config['method'] = 'create';
+
+        $userCatalogues = $this->userRoleRepository->all(['permissions']);
+        $permissions = $this->permissionRepository->all();
+        return view('backend.dashboard.layout', compact(
+            'template',
+            'userCatalogues',
+            'permissions',
+            'config',
+        ));
+    }
+
+    public function updatePermission(Request $request){
+        // $this->authorize('modules', 'user.role.updatePermission');
+        if($this->userRoleService->setPermission($request)){
+            return redirect()->route('user.role.permission')->with('success', 'Cập nhập quyền thành công !');
+        }
+        return redirect()->route('user.role.permission')->with('error', 'Cập nhập quyền thất bại !');
     }
 
 }
